rename: js 파일 내에 폴더로 기능 구분
